Page updated with new flow
There is no tag available for [HoloLens 2 Programming] on our Q&A portal. For now, this will take the user to the AOA page. This needs to be fixed before the build is released

Section needs to be revised with correct support types on Azure portal - Open an Azure tech support ticket

We need a new tag on feedback.azure.com for hololens. which is not available for now. The product Feedback section on this page should be updated and new URL should be added in this section.