Fix correctness issues
https://microsoft-ce-csi.acrolinx.cloud/api/v1/checking/scorecards/d517fa2c-c05b-4307-891e-f8ea4f45c6b7#CORRECTNESS
Line 80: There's a known issue where when the device reaches 18% battery, it will unexpectedly shut down. (A known issue is that when the device reaches 18% battery, it will shut down unexpectedly.)
Line 97: you may have encountered error when sign in to your inbox OneDrive app. (an error when signing in)
Line 122: If you're unable to update Remote Assist to a newer build try the following work around. (If you're unable to update Remote Assist to a newer build, try the following workaround: )
Line 217: A HoloLens 2 device can be configured to automatically login in via (login in via --> log in through)
Line 217: Some users may be required to log-in to the device again when updating a device with a substantially large update, such as a feature update. (Some users may be required to log in to the device again during a substantially large update, such as a feature update.)